pyproject.toml
[project]
name = "nonebot-plugin-shiro-web-console"
version = "0.1.5"
description = "一个用于 NoneBot2 的网页控制台插件，支持通过浏览器查看日志和发送消息"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "luojisama", email = "2534316454@qq.com" },
]
license = { text = "MIT" }
keywords = ["nonebot", "nonebot2", "plugin", "webconsole", "console"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "httpx>=0.23.0",
    "jinja2>=3.1.0",
    "nonebot-plugin-localstore>=0.6.0",
    "pydantic",
]

[project.urls]
"Homepage" = "https://github.com/luojisama/nonebot-plugin-shiro-web-console"
"Bug Tracker" = "https://github.com/luojisama/nonebot-plugin-shiro-web-console/issues"
"Repository" = "https://github.com/luojisama/nonebot-plugin-shiro-web-console"

[tool.nonebot]
plugins = ["nonebot_plugin_shiro_web_console"]
plugin_dirs = ["nonebot_plugin_shiro_web_console"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_shiro_web_console"]

[tool.hatch.build.targets.sdist]
include = [
    "/nonebot_plugin_shiro_web_console",
]
